+++ react-components/src/components/Table/components/default/custom-cell/index.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import { ICellComponent } from '../../../model/public-api';
+
+export function CustomCell({
+  column,
+  row,
+}: React.PropsWithoutRef<ICellComponent<object>>) {
+  const value = row.getValue<unknown>(column.id);
+  return typeof value === 'object' ? JSON.stringify(value) : String(value);
+}